Split large UI feature modules
