+++ apps/ui/src/app/features/realtime_feature.ts
@@ -3,7 +3,6 @@ import { deriveCarSelectionState } from "../car_selection_state";
 import type { RealtimeState, SettingsState, SpectrumState } from "../ui_app_state";
 import type { LocationOption, LoggingStatusPayload } from "../../api/types";
 import type { AdaptedClient } from "../../server_payload";
-import * as I18N from "../../i18n";
 import {
   getClientLocations,
   getLoggingStatus,
@@ -27,6 +26,7 @@ import {
 } from "../views/realtime_sensor_table_view";
 import { createPollingController } from "./polling_controller";
 import { classifyDataFreshness } from "./data_freshness";
+import { createRealtimeSensorState } from "./realtime_sensor_state";
 
 export interface RealtimeFeatureDeps extends FeatureDepsBase {
   realtime: RealtimeState;
@@ -54,11 +54,6 @@ export interface RealtimeFeature {
   refreshLocationOptions(): Promise<void>;
 }
 
-type ActiveCarDisplayState = {
-  text: string;
-  isWarning: boolean;
-};
-
 type CaptureReadinessPayload = NonNullable<LoggingStatusPayload["capture_readiness"]>;
 type CaptureReadinessCheckPayload = CaptureReadinessPayload["checks"][number];
 
@@ -83,123 +78,29 @@ export function createRealtimeFeature(ctx: RealtimeFeatureDeps): RealtimeFeature
     "capture_ready",
   ] as const;
 
-  const SHORTHAND_LOCATION_MAP: Record<string, string> = {
-    "front left": "front_left_wheel",
-    "front right": "front_right_wheel",
-    "rear left": "rear_left_wheel",
-    "rear right": "rear_right_wheel",
-    driver: "driver_seat",
-  };
-
-  function locationLabelForLang(lang: string, code: string): string {
-    return I18N.get(lang, `location.${code}`, { code });
-  }
-
-  function locationLabel(code: string): string {
-    return locationLabelForLang(ctx.getLanguage(), code);
-  }
-
-  function buildLocationOptions(codes: readonly string[]): LocationOption[] {
-    return codes.map((code) => ({ code, label: locationLabel(code) }));
-  }
+  const sensorState = createRealtimeSensorState({
+    realtime,
+    settings,
+    spectrum,
+    getLanguage: ctx.getLanguage,
+    t,
+    formatInt,
+  });
+  const {
+    activeCarDisplayState,
+    assignedClientCount,
+    buildLocationOptions,
+    computeLiveHealth,
+    connectedClients,
+    hasActiveCarSelection,
+    locationCodeForClient,
+    strongestSignal,
+    strongestSignalText,
+  } = sensorState;
 
   function applyLocationCodes(codes: string[]): void {
     realtime.locationCodes = codes.length ? codes : defaultLocationCodes.slice();
     realtime.locationOptions = buildLocationOptions(realtime.locationCodes);
-  }
-
-  function locationCodeForClient(client: AdaptedClient): string {
-    const explicitCode = String(client.location_code || "").trim();
-    if (explicitCode && realtime.locationCodes.includes(explicitCode)) return explicitCode;
-    const name = String(client.name || "").trim();
-    if (!name) return "";
-    const normalizedName = name.toLowerCase().replace(/[_-]+/g, " ").replace(/\s+/g, " ").trim();
-    for (const [token, code] of Object.entries(SHORTHAND_LOCATION_MAP)) {
-      if (normalizedName.includes(token) && realtime.locationCodes.includes(code)) return code;
-    }
-    for (const code of realtime.locationCodes) {
-      const labels = I18N.getForAllLangs(`location.${code}`);
-      if (labels.some((label) => label === name)) return code;
-    }
-    const match = realtime.locationOptions.find((loc) => loc.label === name);
-    return match ? match.code : "";
-  }
-
-  function clientDisplayName(client: AdaptedClient): string {
-    return String(client.name || "").trim() || client.id;
-  }
-
-  function clientLocationText(client: AdaptedClient): string {
-    const code = locationCodeForClient(client);
-    if (!code) {
-      return t("dashboard.sensor_unassigned");
-    }
-    const option = realtime.locationOptions.find((location) => location.code === code);
-    return option?.label ?? locationLabel(code);
-  }
-
-  function connectedClients(): AdaptedClient[] {
-    return realtime.clients.filter((client) => Boolean(client.connected));
-  }
-
-  function assignedClientCount(): number {
-    return realtime.clients.filter((client) => locationCodeForClient(client)).length;
-  }
-
-  function strongestSignal(): { client: AdaptedClient; db: number } | null {
-    let bestClient: AdaptedClient | null = null;
-    let bestDb = Number.NEGATIVE_INFINITY;
-    for (const client of connectedClients()) {
-      const db = spectrum.spectra.clients[client.id]?.strength_metrics?.vibration_strength_db;
-      if (typeof db !== "number" || !Number.isFinite(db)) continue;
-      if (db > bestDb) {
-        bestDb = db;
-        bestClient = client;
-      }
-    }
-    if (!bestClient) {
-      return null;
-    }
-    return {
-      client: bestClient,
-      db: bestDb,
-    };
-  }
-
-  function strongestSignalText(signal = strongestSignal()): string {
-    if (!signal) {
-      return t("dashboard.strongest_signal_none");
-    }
-    const primary = locationCodeForClient(signal.client)
-      ? clientLocationText(signal.client)
-      : clientDisplayName(signal.client);
-    return `${primary} (${formatInt(signal.db)} dB)`;
-  }
-
-  function activeCarDisplayState(): ActiveCarDisplayState {
-    const selection = deriveCarSelectionState(settings);
-    if (selection.kind === "loading") {
-      return {
-        text: t("dashboard.active_car_loading"),
-        isWarning: false,
-      };
-    }
-    if (selection.kind !== "active") {
-      return {
-        text: selection.kind === "no_cars"
-          ? t("dashboard.active_car_none_no_cars")
-          : t("dashboard.active_car_none_blocked"),
-        isWarning: true,
-      };
-    }
-    return {
-      text: selection.car.name,
-      isWarning: false,
-    };
-  }
-
-  function hasActiveCarSelection(): boolean {
-    return deriveCarSelectionState(settings).kind === "active";
   }
 
   function renderActiveCarStat(): void {
@@ -262,13 +163,6 @@ export function createRealtimeFeature(ctx: RealtimeFeatureDeps): RealtimeFeature
     return t("dashboard.data_freshness_stale", { age: ageText });
   }
 
-  type LiveHealth = {
-    variant: "muted" | "ok" | "warn" | "bad";
-    text: string;
-    summary: string;
-    showOverviewPill: boolean;
-  };
-
   type RecordingPanelState = {
     pillVariant: "muted" | "ok" | "warn" | "bad";
     pillText: string;
@@ -313,84 +207,8 @@ export function createRealtimeFeature(ctx: RealtimeFeatureDeps): RealtimeFeature
     }
   }
 
-  function computeLiveHealth(): LiveHealth {
-    if (realtime.loggingStatus.write_error) {
-      return {
-        variant: "bad",
-        text: t("dashboard.health.write_error"),
-        summary: realtime.loggingStatus.write_error,
-        showOverviewPill: true,
-      };
-    }
-    const connected = connectedClients();
-    if (!connected.length) {
-      return {
-        variant: "muted",
-        text: t("dashboard.health.no_signal"),
-        summary: t("dashboard.logging.waiting"),
-        showOverviewPill: true,
-      };
-    }
-    if (!hasActiveCarSelection()) {
-      return {
-        variant: "warn",
-        text: t("dashboard.health.attention"),
-        summary: t("dashboard.logging.active_car_required"),
-        showOverviewPill: true,
-      };
-    }
-    const droppedCount = connected.filter((client) => (client.dropped_frames ?? 0) > 0).length;
-    if (droppedCount > 0) {
-      return {
-        variant: "warn",
-        text: t("dashboard.health.attention"),
-        summary: t("dashboard.logging.frame_loss", { count: formatInt(droppedCount) }),
-        showOverviewPill: true,
-      };
-    }
-    const unassignedConnectedCount = connected.filter((client) => !locationCodeForClient(client)).length;
-    if (unassignedConnectedCount > 0) {
-      return {
-        variant: "warn",
-        text: t("dashboard.health.attention"),
-        summary: t("dashboard.logging.unassigned", { count: formatInt(unassignedConnectedCount) }),
-        showOverviewPill: true,
-      };
-    }
-    const offlineCount = realtime.clients.filter((client) => !client.connected).length;
-    if (offlineCount > 0) {
-      return {
-        variant: "warn",
-        text: t("dashboard.health.attention"),
-        summary: t("dashboard.logging.offline", { count: formatInt(offlineCount) }),
-        showOverviewPill: true,
-      };
-    }
-    const connectedCount = formatInt(connected.length);
-    const assignedCount = formatInt(assignedClientCount());
-    if (realtime.loggingStatus.enabled) {
-      return {
-        variant: "ok",
-        text: t("dashboard.health.recording"),
-        summary: t("dashboard.logging.running", { connected: connectedCount, assigned: assignedCount }),
-        showOverviewPill: false,
-      };
-    }
-    const captureReadiness = realtime.loggingStatus.capture_readiness ?? null;
-    if (captureReadiness && !captureReadiness.is_ready) {
-      return {
-        variant: "warn",
-        text: t("dashboard.health.attention"),
-        summary: captureReadinessSummaryText(captureReadiness),
-        showOverviewPill: true,
-      };
-    }
-    return {
-      variant: "ok",
-      text: t("dashboard.health.ready"),
-      summary: "",
-      showOverviewPill: false,
-    };
+  function computeCurrentLiveHealth() {
+    return computeLiveHealth(captureReadinessSummaryText);
   }
 
   function renderLiveSensorRoster(): void {
@@ -418,7 +236,7 @@ export function createRealtimeFeature(ctx: RealtimeFeatureDeps): RealtimeFeature
   }
 
   function renderLiveHealth(): void {
-    const health = computeLiveHealth();
+    const health = computeCurrentLiveHealth();
     setDashboardPillState(els.liveRunHealth, health.variant, health.text, { hidden: !health.showOverviewPill });
     setPillState(els.shellLiveStatus, health.variant, health.text);
   }
@@ -899,7 +717,7 @@ export function createRealtimeFeature(ctx: RealtimeFeatureDeps): RealtimeFeature
     const hasActiveCar = hasActiveCarSelection();
     const connectedCount = formatInt(connectedClients().length);
     const assignedCount = formatInt(assignedClientCount());
-    const liveHealth = computeLiveHealth();
+    const liveHealth = computeCurrentLiveHealth();
     const runIdText = recordingRunIdText(status);
     const elapsedText = on ? formatElapsed(status.start_time_utc) : "--";
     const samplesText = formatInt(status.samples_written ?? 0);
